src/video-intelligence-streamlit/requirements.txt
streamlit
pytubefix
google-auth
google-cloud-aiplatform